Moving test files other things for obspy.core.inventory
