+++ obspy/core/tests/data/AU.MEEK.seed
