+++ obspy/core/tests/data/AU.MEEK.xml
