+++ obspy/core/tests/data/BW_GR_misc.xml.gz
