+++ obspy/core/tests/data/BW_RJOB.xml
