+++ obspy/core/tests/data/IRIS_single_channel_with_response.seed
