+++ obspy/core/tests/data/IU_ANMO_00_BHZ.xml
