+++ obspy/core/tests/data/IU_ANMO_BH.xml
