+++ obspy/core/tests/data/Modified_IRIS_response_level_station.xml
