+++ obspy/core/tests/data/TM.SKLT..BHZ_faulty_response.xml
