+++ obspy/core/tests/data/XM.05.xml
