+++ obspy/core/tests/data/full_network_field_station.xml
