+++ obspy/core/tests/data/full_random_stationxml.xml
