+++ obspy/core/tests/data/full_station_field_station.xml
