+++ obspy/core/tests/data/minimal_station.xml
