+++ obspy/core/tests/data/minimal_with_non_obspy_module_and_sender_tags_station.xml
