+++ obspy/core/tests/data/stationxml_with_availability.xml
