+++ obspy/core/tests/images/channel_response.png
