+++ obspy/core/tests/images/inventory_location1.png
