+++ obspy/core/tests/images/inventory_location2.png
