+++ obspy/core/tests/images/inventory_location3.png
