+++ obspy/core/tests/images/inventory_response.png
