+++ obspy/core/tests/images/network_location1.png
