+++ obspy/core/tests/images/network_location2.png
