+++ obspy/core/tests/images/network_location3.png
