+++ obspy/core/tests/images/network_response.png
