+++ obspy/core/tests/images/response_response.png
