+++ obspy/core/tests/images/station_response.png
