+++ obspy/core/tests/test_network.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
-Test suite for the inventory class.
+Test suite for the network class.
 
 :copyright:
     Lion Krischer (krischer@geophysik.uni-muenchen.de), 2013
@@ -23,7 +23,7 @@ from obspy import UTCDateTime, read_inventory
 from obspy.core.util.base import getBasemapVersion
 from obspy.core.util.decorator import skipIf
 from obspy.core.util.testing import ImageComparison, getMatplotlibVersion
-from obspy.station import Channel, Inventory, Network, Response, Station
+from obspy.core.inventory import Channel, Network, Response, Station
 
 # checking for matplotlib/basemap
 try:
@@ -36,13 +36,13 @@ except ImportError:
     HAS_BASEMAP = False
 
 
-MATPLOTLIB_VERSION = getMatplotlibVersion()
 BASEMAP_VERSION = getBasemapVersion()
+MATPLOTLIB_VERSION = getMatplotlibVersion()
 
 
-class InventoryTestCase(unittest.TestCase):
+class NetworkTestCase(unittest.TestCase):
     """
-    Tests the for :class:`~obspy.station.inventory.Inventory` class.
+    Tests the for :class:`~obspy.station.network.Network` class.
     """
     def setUp(self):
         self.image_dir = os.path.join(os.path.dirname(__file__), 'images')
@@ -51,17 +51,6 @@ class InventoryTestCase(unittest.TestCase):
 
     def tearDown(self):
         np.seterr(**self.nperr)
-
-    def test_initialization(self):
-        """
-        Some simple sanity tests.
-        """
-        dt = UTCDateTime()
-        inv = Inventory(source="TEST", networks=[])
-        # If no time is given, the creation time should be set to the current
-        # time. Use a large offset for potentially slow computers and test
-        # runs.
-        self.assertTrue(inv.created - dt <= 10.0)
 
     def test_get_response(self):
         responseN1S1 = Response('RESPN1S1')
@@ -97,24 +86,22 @@ class InventoryTestCase(unittest.TestCase):
                              latitude=0.0,
                              longitude=0.0,
                              elevation=0.0,
-                             channels=channelsN1S2)]
-        stations2 = [Station(code='N2S1',
+                             channels=channelsN1S2),
+                     Station(code='N2S1',
                              latitude=0.0,
                              longitude=0.0,
                              elevation=0.0,
                              channels=channelsN2S1)]
-        networks = [Network('N1', stations=stations1),
-                    Network('N2', stations=stations2)]
-        inv = Inventory(networks=networks, source='TEST')
+        network = Network('N1', stations=stations1)
 
-        response = inv.get_response('N1.N1S1..BHZ',
-                                    UTCDateTime('2010-01-01T12:00'))
+        response = network.get_response('N1.N1S1..BHZ',
+                                        UTCDateTime('2010-01-01T12:00'))
         self.assertEqual(response, responseN1S1)
-        response = inv.get_response('N1.N1S2..BHZ',
-                                    UTCDateTime('2010-01-01T12:00'))
+        response = network.get_response('N1.N1S2..BHZ',
+                                        UTCDateTime('2010-01-01T12:00'))
         self.assertEqual(response, responseN1S2)
-        response = inv.get_response('N2.N2S1..BHZ',
-                                    UTCDateTime('2010-01-01T12:00'))
+        response = network.get_response('N1.N2S1..BHZ',
+                                        UTCDateTime('2010-01-01T12:00'))
         self.assertEqual(response, responseN2S1)
 
     def test_get_coordinates(self):
@@ -137,55 +124,53 @@ class InventoryTestCase(unittest.TestCase):
                             longitude=0.0,
                             elevation=0.0,
                             channels=channels)]
-        networks = [Network('BW', stations=stations)]
-        inv = Inventory(networks=networks, source='TEST')
+        network = Network('BW', stations=stations)
         # 1
-        coordinates = inv.get_coordinates('BW.RJOB..EHZ',
-                                          UTCDateTime('2010-01-01T12:00'))
+        coordinates = network.get_coordinates('BW.RJOB..EHZ',
+                                              UTCDateTime('2010-01-01T12:00'))
         self.assertEqual(sorted(coordinates.items()), sorted(expected.items()))
         # 2 - without datetime
-        coordinates = inv.get_coordinates('BW.RJOB..EHZ')
+        coordinates = network.get_coordinates('BW.RJOB..EHZ')
         self.assertEqual(sorted(coordinates.items()), sorted(expected.items()))
         # 3 - unknown SEED ID should raise exception
-        self.assertRaises(Exception, inv.get_coordinates, 'BW.RJOB..XXX')
+        self.assertRaises(Exception, network.get_coordinates, 'BW.RJOB..XXX')
 
     @skipIf(not HAS_BASEMAP, 'basemap not installed')
     def test_location_plot_global(self):
         """
-        Tests the inventory location preview plot, default parameters.
+        Tests the network location preview plot, default parameters.
         """
-        inv = read_inventory()
+        net = read_inventory()[0]
         reltol = 1.0
         # Coordinate lines might be slightly off, depending on the basemap
         # version.
         if BASEMAP_VERSION < [1, 0, 7]:
             reltol = 3.0
-        with ImageComparison(self.image_dir, "inventory_location1.png",
+        with ImageComparison(self.image_dir, "network_location1.png",
                              reltol=reltol) as ic:
             rcParams['savefig.dpi'] = 72
-            inv.plot(outfile=ic.name)
+            net.plot(outfile=ic.name)
 
     @skipIf(not HAS_BASEMAP, 'basemap not installed')
     def test_location_plot_ortho(self):
         """
-        Tests the inventory location preview plot, ortho projection, some
+        Tests the network location preview plot, ortho projection, some
         non-default parameters.
         """
-        inv = read_inventory()
-        with ImageComparison(self.image_dir, "inventory_location2.png") as ic:
+        net = read_inventory()[0]
+        with ImageComparison(self.image_dir, "network_location2.png") as ic:
             rcParams['savefig.dpi'] = 72
-            inv.plot(projection="ortho", resolution="c",
-                     continent_fill_color="0.3", marker="d",
-                     label=False, outfile=ic.name, colormap="hsv",
-                     color_per_network=True)
+            net.plot(projection="ortho", resolution="c",
+                     continent_fill_color="0.5", marker="d",
+                     color="yellow", label=False, outfile=ic.name)
 
     @skipIf(not HAS_BASEMAP, 'basemap not installed')
     def test_location_plot_local(self):
         """
-        Tests the inventory location preview plot, local projection, some more
+        Tests the network location preview plot, local projection, some more
         non-default parameters.
         """
-        inv = read_inventory()
+        net = read_inventory()[0]
         # Coordinate lines might be slightly off, depending on the basemap
         # version.
         reltol = 2.0
@@ -193,11 +178,10 @@ class InventoryTestCase(unittest.TestCase):
         # tolerance must be much higher.
         if BASEMAP_VERSION < [1, 0, 4]:
             reltol = 100.0
-        with ImageComparison(self.image_dir, "inventory_location3.png",
+        with ImageComparison(self.image_dir, "network_location3.png",
                              reltol=reltol) as ic:
             rcParams['savefig.dpi'] = 72
-            inv.plot(projection="local", resolution="i", size=20**2,
-                     color_per_network={"GR": "b", "BW": "green"},
+            net.plot(projection="local", resolution="i", size=13**2,
                      outfile=ic.name)
 
     def test_response_plot(self):
@@ -210,70 +194,19 @@ class InventoryTestCase(unittest.TestCase):
         if [1, 4, 0] <= MATPLOTLIB_VERSION <= [1, 5, 0]:
             reltol = 2.0
 
-        inv = read_inventory()
+        net = read_inventory()[0]
         t = UTCDateTime(2008, 7, 1)
         with warnings.catch_warnings(record=True):
             warnings.simplefilter("ignore")
-            with ImageComparison(self.image_dir, "inventory_response.png",
+            with ImageComparison(self.image_dir, "network_response.png",
                                  reltol=reltol) as ic:
                 rcParams['savefig.dpi'] = 72
-                inv.plot_response(0.01, output="ACC", channel="*N",
-                                  station="[WR]*", time=t, outfile=ic.name)
-
-    def test_inventory_merging_metadata_update(self):
-        """
-        Tests the metadata update during merging of inventory objects.
-        """
-        inv_1 = read_inventory()
-        inv_2 = read_inventory()
-
-        inv_1 += inv_2
-
-        self.assertEqual(inv_1.source, inv_2.source)
-        self.assertEqual(inv_1.sender, inv_2.sender)
-        self.assertTrue("ObsPy" in inv_1.module)
-        self.assertTrue("obspy.org" in inv_1.module_uri)
-        self.assertTrue((UTCDateTime() - inv_1.created) < 5)
-
-        # Now a more advanced case.
-        inv_1 = read_inventory()
-        inv_2 = read_inventory()
-
-        inv_1.source = "B"
-        inv_2.source = "A"
-
-        inv_1.sender = "Random"
-        inv_2.sender = "String"
-
-        inv_1 += inv_2
-
-        self.assertEqual(inv_1.source, "A,B")
-        self.assertEqual(inv_1.sender, "Random,String")
-        self.assertTrue("ObsPy" in inv_1.module)
-        self.assertTrue("obspy.org" in inv_1.module_uri)
-        self.assertTrue((UTCDateTime() - inv_1.created) < 5)
-
-        # One more. Containing a couple of Nones.
-        inv_1 = read_inventory()
-        inv_2 = read_inventory()
-
-        inv_1.source = None
-        inv_2.source = "A"
-
-        inv_1.sender = "Random"
-        inv_2.sender = None
-
-        inv_1 += inv_2
-
-        self.assertEqual(inv_1.source, "A")
-        self.assertEqual(inv_1.sender, "Random")
-        self.assertTrue("ObsPy" in inv_1.module)
-        self.assertTrue("obspy.org" in inv_1.module_uri)
-        self.assertTrue((UTCDateTime() - inv_1.created) < 5)
+                net.plot_response(0.002, output="DISP", channel="B*E",
+                                  time=t, outfile=ic.name)
 
 
 def suite():
-    return unittest.makeSuite(InventoryTestCase, 'test')
+    return unittest.makeSuite(NetworkTestCase, 'test')
 
 
 if __name__ == '__main__':
