+++ obspy/core/tests/test_station.py
@@ -20,8 +20,8 @@ import warnings
 import numpy as np
 from matplotlib import rcParams
 
+from obspy import read_inventory
 from obspy.core.util.testing import ImageComparison, getMatplotlibVersion
-from obspy.station import read_inventory
 
 
 MATPLOTLIB_VERSION = getMatplotlibVersion()
