+++ obspy/core/tests/test_stationxml.py
